# Documentation/openapi.yaml
openapi: 3.0.3
info:
  title: AIS Clinic API
  version: 1.0.0
  description: OpenAPI-описание ключевых эндпоинтов АИС "Клиника"
servers:
  - url: https://api.clinic.local
    description: Production
  - url: http://localhost:8000
    description: Local development
...
security:
  - bearerAuth: []
paths:
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Аутентификация пользователя
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/LoginFormRequest'
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Получить текущего пользователя по JWT
      responses:
        '200':
          description: Текущий пользователь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Пользователь неактивен или не имеет доступа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/doctors:
    post:
      tags: [Doctors]
      summary: Создание врача (кадровый контур)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DoctorCreateRequest'
      responses:
        '201':
          description: Врач создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DoctorListItem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/doctors/{doctor_id}/employment:
    patch:
      tags: [Doctors]
      summary: Обновить кадровые параметры врача (оклад/премия/статус)
      parameters:
        - in: path
          name: doctor_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DoctorEmploymentUpdateRequest'
      responses:
        '200':
          description: Параметры врача обновлены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DoctorListItem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /api/v1/appointments:
    post:
      tags: [Appointments]
      summary: Запись пациента на прием
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentCreateRequest'
      responses:
        '201':
          description: Запись создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
    get:
      tags: [Appointments]
      summary: Получить список записей
      parameters:
        - in: query
          name: doctor_id
          schema:
            type: string
            format: uuid
        - in: query
          name: patient_id
          schema:
            type: string
            format: uuid
        - in: query
          name: status
          schema:
            type: string
            enum: [planned, confirmed, cancelled, completed, no_show]
        - in: query
          name: skip
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: Список записей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/appointments/{appointment_id}/cancel:
    patch:
      tags: [Appointments]
      summary: Отмена записи
      parameters:
        - in: path
          name: appointment_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentCancelRequest'
      responses:
        '200':
          description: Запись отменена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/appointments/{appointment_id}/reschedule:
    patch:
      tags: [Appointments]
      summary: Перенос записи на другой слот
      parameters:
        - in: path
          name: appointment_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentRescheduleRequest'
      responses:
        '200':
          description: Запись перенесена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /api/v1/labs/orders:
    post:
      tags: [Laboratory]
      summary: Создать лабораторное направление по визиту
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LabOrderCreateRequest'
      responses:
        '201':
          description: Направление создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabOrderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/labs/orders/{order_id}/items:
    post:
      tags: [Laboratory]
      summary: Добавить тесты в лабораторное направление
      parameters:
        - in: path
          name: order_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LabOrderItemsCreateRequest'
      responses:
        '201':
          description: Тесты добавлены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabOrderItemsCreateResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/labs/results:
    post:
      tags: [Laboratory]
      summary: Внести результат анализа
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LabResultCreateRequest'
      responses:
        '201':
          description: Результат внесен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabResultResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
    get:
      tags: [Laboratory]
      summary: Список лабораторных результатов
      parameters:
        - in: query
          name: patient_id
          schema:
            type: string
            format: uuid
        - in: query
          name: order_id
          schema:
            type: string
            format: uuid
        - in: query
          name: skip
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: Список результатов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LabResultListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/v1/reports/appointments:
    get:
      tags: [Reports]
      summary: Статистика записей за период
      parameters:
        - in: query
          name: date_from
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: date_to
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Сводка по статусам записей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentsStatsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ValidationError:
      description: Ошибка валидации входных данных
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Требуется аутентификация
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Недостаточно прав
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Конфликт состояния
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    LoginFormRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 100
        password:
          type: string
          minLength: 1
          maxLength: 128
...
    TokenResponse:
      type: object
      required: [access_token, token_type, expires_in]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          enum: [bearer]
        expires_in:
          type: integer
          minimum: 1
    CurrentUserResponse:
      type: object
      required: [id, username, email, role, is_active]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, doctor, registrar, patient]
        is_active:
          type: boolean
    DoctorListItem:
      type: object
      required: [id, full_name, license_number, experience_years, specializations, monthly_salary, bonus_percent, is_active]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        last_name:
          type: string
          nullable: true
        first_name:
          type: string
          nullable: true
        middle_name:
          type: string
          nullable: true
        full_name:
          type: string
        license_number:
          type: string
        experience_years:
          type: integer
          minimum: 0
        specializations:
          type: array
          items:
            type: string
        monthly_salary:
          type: number
          minimum: 0
        bonus_percent:
          type: number
          minimum: 0
          maximum: 100
        is_active:
          type: boolean
...
    DoctorCreateRequest:
      type: object
      required: [username, email, password, last_name, first_name, license_number, experience_years]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 100
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 128
        last_name:
          type: string
          minLength: 1
          maxLength: 100
        first_name:
          type: string
          minLength: 1
          maxLength: 100
        middle_name:
          type: string
          maxLength: 100
        license_number:
          type: string
          minLength: 3
          maxLength: 50
        experience_years:
          type: integer
          minimum: 0
          maximum: 60
        specialty_ids:
          type: array
          items:
            type: integer
            minimum: 1
          maxItems: 10
        monthly_salary:
          type: number
          minimum: 0
          maximum: 10000000
          default: 0
        bonus_percent:
          type: number
          minimum: 0
          maximum: 100
          default: 0
        is_active:
          type: boolean
          default: true
    DoctorEmploymentUpdateRequest:
      type: object
      properties:
        monthly_salary:
          type: number
          minimum: 0
          maximum: 10000000
        bonus_percent:
          type: number
          minimum: 0
          maximum: 100
        is_active:
          type: boolean
...
    AppointmentCreateRequest:
      type: object
      required: [patient_id, doctor_id, slot_id]
      properties:
        patient_id:
          type: string
          format: uuid
        doctor_id:
          type: string
          format: uuid
        slot_id:
          type: string
          format: uuid
        reason:
          type: string
          maxLength: 2000
    AppointmentCancelRequest:
      type: object
      properties:
        reason:
          type: string
          maxLength: 500
    AppointmentRescheduleRequest:
      type: object
      required: [slot_id]
      properties:
        slot_id:
          type: string
          format: uuid
    AppointmentResponse:
      type: object
      required: [id, patient_id, doctor_id, slot_id, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        patient_id:
          type: string
          format: uuid
        patient_full_name:
          type: string
          nullable: true
        doctor_id:
          type: string
          format: uuid
        doctor_full_name:
          type: string
          nullable: true
        slot_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [planned, confirmed, cancelled, completed, no_show]
        created_at:
          type: string
          format: date-time
    AppointmentListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AppointmentResponse'
        total:
          type: integer
          minimum: 0
...
    LabOrderCreateRequest:
      type: object
      required: [visit_id]
      properties:
        visit_id:
          type: string
          format: uuid
    LabOrderResponse:
      type: object
      required: [id, visit_id, patient_id, ordered_by_doctor_id, status, ordered_at]
      properties:
        id:
          type: string
          format: uuid
        visit_id:
          type: string
          format: uuid
        patient_id:
          type: string
          format: uuid
        ordered_by_doctor_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [created, in_progress, completed, cancelled]
        ordered_at:
          type: string
          format: date-time
    LabOrderItemCreateItem:
      type: object
      required: [test_type_id]
      properties:
        test_type_id:
          type: string
          format: uuid
        priority:
          type: string
          enum: [normal, urgent, stat]
          default: normal
    LabOrderItemsCreateRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/LabOrderItemCreateItem'
    LabOrderItemResponse:
      type: object
      required: [id, lab_order_id, test_type_id, priority]
      properties:
        id:
          type: string
          format: uuid
        lab_order_id:
          type: string
          format: uuid
        test_type_id:
          type: string
          format: uuid
        priority:
          type: string
          enum: [normal, urgent, stat]
    LabOrderItemsCreateResponse:
      type: object
      required: [lab_order_id, created, items]
      properties:
        lab_order_id:
          type: string
          format: uuid
        created:
          type: integer
          minimum: 0
        items:
          type: array
          items:
            $ref: '#/components/schemas/LabOrderItemResponse'
    LabResultCreateRequest:
      type: object
      required: [lab_order_item_id, flag, validated_by_doctor_id]
      properties:
        lab_order_item_id:
          type: string
          format: uuid
        value_text:
          type: string
          maxLength: 100
          nullable: true
        value_numeric:
          type: number
          format: double
          nullable: true
        unit:
          type: string
          maxLength: 30
        reference_range:
          type: string
          maxLength: 50
        flag:
          type: string
          enum: [low, normal, high, critical]
        validated_by_doctor_id:
          type: string
          format: uuid
      oneOf:
        - required: [value_text]
        - required: [value_numeric]
    LabResultResponse:
      type: object
      required: [id, lab_order_item_id, lab_order_id, patient_id, resulted_at]
      properties:
        id:
          type: string
          format: uuid
        lab_order_item_id:
          type: string
          format: uuid
        lab_order_id:
          type: string
          format: uuid
        patient_id:
          type: string
          format: uuid
        value_text:
          type: string
          nullable: true
        value_numeric:
          type: number
          format: double
          nullable: true
        unit:
          type: string
          nullable: true
        reference_range:
          type: string
          nullable: true
        flag:
          type: string
          enum: [low, normal, high, critical]
          nullable: true
        resulted_at:
          type: string
          format: date-time
        validated_by_doctor_id:
          type: string
          format: uuid
          nullable: true
    LabResultListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/LabResultResponse'
        total:
          type: integer
          minimum: 0
...
    AppointmentsStatsResponse:
      type: object
      required: [date_from, date_to, total, planned, confirmed, completed, cancelled, no_show]
      properties:
        date_from:
          type: string
          format: date
        date_to:
          type: string
          format: date
        total:
          type: integer
          minimum: 0
        planned:
          type: integer
          minimum: 0
        confirmed:
          type: integer
          minimum: 0
        completed:
          type: integer
          minimum: 0
        cancelled:
          type: integer
          minimum: 0
        no_show:
          type: integer
          minimum: 0
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  issue:
                    type: string
            request_id:
              type: string
              format: uuid